Move Rule to its own file
