+++ src/LaunchDarkly.Client/FeatureFlag.cs
@@ -39,6 +39,7 @@ namespace LaunchDarkly.Client
             Deleted = deleted;
         }
 
+
         public FeatureFlag()
         {
         }
@@ -187,30 +188,6 @@ namespace LaunchDarkly.Client
         {
             Variation = variation;
             Weight = weight;
-        }
-    }
-
-    public class Rule : VariationOrRollout
-    {
-        internal List<Clause> Clauses { get; }
-
-        [JsonConstructor]
-        public Rule(int? variation, Rollout rollout, List<Clause> clauses) : base(variation, rollout)
-        {
-            Clauses = clauses;
-        }
-
-        internal bool MatchesUser(User user)
-        {
-            foreach (var c in Clauses)
-            {
-                if (!c.MatchesUser(user))
-                {
-                    return false;
-                }
-
-            }
-            return true;
         }
     }
